drm/i915/pf: Save and restore VFs state during S2idle/S3/S4
During suspend and hibernation of the host, the VFs that are passed to
the VM are not aware of the suspend/hibernation and therefore cannot
prepare for it and then reinitialize, just as the driver on the host does.
However, we can use the GuC state saving mechanism (created for cold
migration) to save the GuC state for VFs at the suspend, and then restore
it on the resume path.
Let's save and restore VFs GuC state, and additionally let's reinitialize
some VFs PCI configs and re-enable VFs interrupts for VFs, to restore full VFs
functionality on VM, after S2idle/S3/S4.

v2: check for vfpdev to make KW happy (michal)
v3: include a variable storing GuC state size (tomaszli)

Signed-off-by: Piotr Piórkowski <piotr.piorkowski@intel.com> #v1 #v2
Signed-off-by: Tomasz Lis <tomasz.lis@intel.com> #v3
Acked-by: Michal Wajdeczko <michal.wajdeczko@intel.com> #v2
Acked-by: Michal Winiarski <michal.winiarski@intel.com> #v3
Signed-off-by: Zawawi, Muhammad Zul Husni <muhammad.zul.husni.zawawi@intel.com>