drm/i915/pf: Skip VF save/restore on S2idle/S3/S4 if it can reinitialize itself
When changing the host state to S2idle/S3/S4, if we want the VF to
survive without errors, we need to save its state and restore it on
resume.
However, if the VF has a driver that can reinitialize after suspend,
we don't have to do anything, just skip.
We can safely assume that if VF has the same driver on the host as PF,
it will survive PM events.
Let's add a condition that will check if we can skip VF save/restore.

v2: check for vfpdev to make KW happy (michal)
    fix leaked vfpdev reference (michal)

Signed-off-by: Piotr Piórkowski <piotr.piorkowski@intel.com>
Acked-by: Michal Wajdeczko <michal.wajdeczko@intel.com>
Signed-off-by: Zawawi, Muhammad Zul Husni <muhammad.zul.husni.zawawi@intel.com>