drm/i915/iov: Introduce VFs shadow copy of GGTT on PF
Due to problems accessing GGTT after through BAR0, we need to have
an alternative way to GGTT save/restore GGTT during migration.
The solution is to keep track of all GGTT changes that VFs make and
store them on PF so that we can do a GGTT save/restore based on it
during the migration.
Add shadow copy of GGTT to storing up-to-date VFs GGTT on PF.

HSDES: 22018453856

Signed-off-by: Piotr Piórkowski <piotr.piorkowski@intel.com>
Signed-off-by: Zawawi, Muhammad Zul Husni <muhammad.zul.husni.zawawi@intel.com>